feat: new workspace home and actions hook
